fstack: External (text) data support
It's sometimes useful to sync uftrace tracing data with some external
data/event.  To do that, add one more data source using a text file
containing timestamp and message for each line.

The text file is 'extern.dat' in the uftrace.data directory.  The
timestamp should have same format as 'time' field in the replay.  The
rest of the line will be treated as a message.  Line started with '#'
character will be treated as comments.

Signed-off-by: Namhyung Kim <namhyung@gmail.com>