mcount: Handle finish trigger only in entry/exit
It's subtle that finish can be triggered by signal so that it can see in
the middle of mcount routine.  Howevery we can delay the processing to
start of mcount_entry() and end of mcount_exit() to prevent possible
races in the middle of the routines.  Especially the exit routines need
to know where to return and it requires rstack to be remained during the
routine.  Same for plthook_entry/exit().

But it's different for cygprof and xray routines since their _exit() can
be called regardless entry() succeeded or not so no need to hook return
addresses.

This effectively reverts commit 368b2b4d ("mcount: Make
mcount/plthook_exit() return 0 when trace finished").

Signed-off-by: Namhyung Kim <namhyung@gmail.com>