mcount: Defer finish trigger in mcount_exit()
It's subtle that finish can be triggered by signal so that it can see in
the middle of mcount routine.  Howevery we can delay the processing to
the end when it sees the trigger in mcount_exit().  The mcount_exit()
needs to know where to return and it requires rstack to be remained
during the routine.  Once it figured out the return address it can
process the finish trigger to free the rstacks.  Same for plthook_exit().

One subtlety is tail-call, which needs to return to itself from the
mcount_exit().  However after rstack is destroyed it should not return
to itself.  To know the real return address, save the pointer where
return address is saved, and reload the pointer.

It's different for cygprof and xray routines since their _exit() can be
called regardless entry() succeeded or not so no need to hook return
addresses.

This effectively reverts commit 368b2b4d ("mcount: Make mcount/plthook_
exit() return 0 when trace finished").

Signed-off-by: Namhyung Kim <namhyung@gmail.com>